Move token module back temporarily
