+++ plt/plt-scheduler/src/lib.rs
@@ -5,6 +5,7 @@ mod locks;
 pub mod protocol_level_tokens;
 pub mod queries;
 pub mod scheduler;
+mod token_module;
 mod transaction_execution;
 
-pub use protocol_level_tokens::token_module::TOKEN_MODULE_REF;
+pub use token_module::TOKEN_MODULE_REF;
