+++ plt/plt-scheduler/src/protocol_level_tokens.rs
@@ -1,3 +1,2 @@
 pub mod p11;
 pub mod p9;
-pub mod token_module;
