+++ plt/plt-scheduler/src/token_module/module/initialize.rs
@@ -1,8 +1,6 @@
 use crate::block_state_polymorph::token::TokenPXRefMut;
-use crate::protocol_level_tokens::token_module::errors::{
-    MintWouldOverflowError, TokenAmountDecimalsMismatchError,
-};
-use crate::protocol_level_tokens::token_module::{balance_operations, util};
+use crate::token_module::errors::{MintWouldOverflowError, TokenAmountDecimalsMismatchError};
+use crate::token_module::{balance_operations, util};
 use concordium_base::common::cbor::CborSerializationError;
 use concordium_base::protocol_level_tokens::{TokenAdminRole, TokenModuleInitializationParameters};
 use plt_block_state::entity::accounts::Accounts;
