+++ plt/plt-scheduler/src/token_module/module/update.rs
@@ -1,8 +1,8 @@
 use crate::block_state_polymorph::token::{TokenPXRef, TokenPXRefMut};
-use crate::protocol_level_tokens::token_module::errors::{
+use crate::token_module::errors::{
     InsufficientBalanceError, MintWouldOverflowError, TokenAmountDecimalsMismatchError,
 };
-use crate::protocol_level_tokens::token_module::{balance_operations, util};
+use crate::token_module::{balance_operations, util};
 use crate::transaction_execution::{OutOfEnergyError, TransactionExecution};
 use concordium_base::base::Energy;
 use concordium_base::contracts_common::AccountAddress;
